refactor: consolidate booking service into internal monorepo structure
Move booking service from standalone microservice to internal/ directory
to align with monorepo architecture. This consolidation includes:

Structural changes:
- Rename adapters/ to infrastructure/ for consistency with other services
- Rename http/ to interface/ to match naming conventions
- Move from booking/internal/ to internal/booking/
- Move integration tests from booking/test/ to test/integration/booking/
- Remove standalone go.mod/go.sum files

Architecture improvements:
- Fix incorrect import path (internal/booking/internal/domain -> internal/booking/domain)
- Move payment constants from infrastructure to ports layer
- Remove infrastructure imports from application layer
- Ensure proper hexagonal architecture compliance

This refactoring maintains all functionality while improving architectural
consistency across the codebase.

+++ backend/internal/catalog/infrastructure/stripe/price/deactivate_reactivate_prices_test.go
@@ -52,7 +52,7 @@ func TestDeactivatePrices(t *testing.T) {
 		for i := 0; i < 3; i++ {
 			createReq := domain.CreateStripePriceRequest{
 				ProductID: productID,
-				Amount:    1000 + (i * 100), // Different amounts
+				Amount:    int64(1000 + (i * 100)), // Different amounts
 				Currency:  "USD",
 				Interval:  "month",
 				Active:    true,
@@ -141,7 +141,7 @@ func TestReactivatePrices(t *testing.T) {
 		for i := 0; i < 3; i++ {
 			createReq := domain.CreateStripePriceRequest{
 				ProductID: productID,
-				Amount:    1500 + (i * 200), // Different amounts
+				Amount:    int64(1500 + (i * 200)), // Different amounts
 				Currency:  "USD",
 				Interval:  "month",
 				Active:    true,
@@ -199,7 +199,7 @@ func TestPriceActivationWorkflow(t *testing.T) {
 		for i := 0; i < 2; i++ {
 			createReq := domain.CreateStripePriceRequest{
 				ProductID: productID,
-				Amount:    1000 + (i * 500),
+				Amount:    int64(1000 + (i * 500)),
 				Currency:  "USD",
 				Interval:  "month",
 				Active:    true,
@@ -245,3 +245,4 @@ func TestPriceActivationWorkflow(t *testing.T) {
 		}
 	})
 }
+
